microservice/requirements.txt
fastapi
uvicorn[standard]
torch
torchvision
pillow
numpy
opencv-python-headless
python-multipart
